Add Hash To Store Body of Section
Add in a hash to store the body of each section's parts using the title
of that part of the sections as the key. Since now we can do sorting by
the keys allowing to loop thought each part, of the section, with
similar part names being together and add them to the database.